final commit, ready for merge
