+++ frontend/src/components/register/index.jsx
@@ -3,26 +3,19 @@ import { Redirect } from "react-router-dom";
 import "./Register.scss";
 import { Link } from 'react-router-dom';
 import { Input, InputPassword, ButtonSubmit } from "../../shared/components/";
-import { GET, POST } from './RegisterAPI.jsx';
+import { POST } from "../../shared/services/requests";
 
 const Register = () => {
   const [name, setName] = useState(''),
         [email, setEmail] = useState(''),
         [pwd, setPwd] = useState(''),
-        [pwdConf, setPwdConf] = useState('');
-        // [toVerify, setToVerify] = useState(false);
+        [pwdConf, setPwdConf] = useState(''),
+        [redirectVerify, setRedirectVerify] = useState(false);
 
   const getName = (e) => setName(e),
         getEmail = (e) => setEmail(e),
         getPwd = (e) => setPwd(e),
         getPwdConf = (e) => setPwdConf(e);
-        
-  // Post data to backend, url needs to be set up in RegisterApi      
-  // const getData = async(apiEndpoint, data) => {
-  //   const { data: items } = await POST(apiEndpoint, data);
-  //   if (items) return console.log(items);
-  //   if (!items) return console.log('error');
-  // }
   
   const submit = () => {
     const data = {
@@ -32,24 +25,15 @@ const Register = () => {
       password_confirmation: pwdConf
     }
     
-    POST('register', data).then(console.log(data))
-    //token exists in data.data.token here, i think. can redirect to logged in page
-
-
-    return <Redirect to="/verify" />;
-
-    // Post data
-    // getData('ApiEndpoint', data);
+    POST('register', data).then(data => {
+      localStorage.setItem('token', data.data.token)
+      setRedirectVerify(true);
+    });
   };
 
-  // const redirect = () => {
-  //   // if (toVerify === true) {
-  //   //   return <Redirect to="/verify" />;
-  //   // }
-  //   console.log("redirect");
-  //   return <Redirect to="/verify" />;
-  // }
-
+  if(redirectVerify)
+    return <Redirect to="/verify" />;
+  
   return (
     <>
       <h1 className="register-title">Sign Up Now</h1>
